chore: Factored out loading logic.
