+++ blocks/models/huggingface/base.py
@@ -1,6 +1,5 @@
 import os
-from threading import local
-from typing import Callable, List, Optional, Tuple, Union
+from typing import List, Optional
 
 import pandas as pd
 import torch
@@ -8,15 +7,7 @@ from blocks.models.base import Model
 from configs.constants import Const
 from datasets import ClassLabel, Dataset, Features, Value
 from sklearn.model_selection import train_test_split
-from transformers import (
-    AutoModelForSequenceClassification,
-    PreTrainedModel,
-    PreTrainedTokenizerBase,
-    Trainer,
-    TrainingArguments,
-    enable_full_determinism,
-    pipeline,
-)
+from transformers import enable_full_determinism
 from type import DataType, Evaluators, HuggingfaceConfig, LoadOrigin, PredsWithProbs
 from utils.env_interface import get_env
 
@@ -24,129 +15,14 @@ from .infer import run_inference
 from .train import run_training
 import textwrap
 from utils.printing import PrintFormats
-
-from transformers import (
-    AutoModelForSequenceClassification,
-    AutoTokenizer,
-    DataCollatorWithPadding,
-    PreTrainedModel,
-    Trainer,
-    TrainingArguments,
-)
+from .loading import safe_load, determine_load_order, get_paths
 
 device = 0 if torch.cuda.is_available() else -1
-
-
-def multi_line_print(text: str) -> None:
-    print(
-        textwrap.fill(
-            text,
-            initial_indent="    ┃  ├── ",
-            subsequent_indent="    ┃  │   ",
-            width=100,
-        )
-    )
-
-
-def safe_load_training_pipeline(
-    module: str, config: HuggingfaceConfig
-) -> Tuple[Optional[PreTrainedModel], Optional[PreTrainedTokenizerBase]]:
-    try:
-        model = AutoModelForSequenceClassification.from_pretrained(
-            module, num_labels=config.num_classes
-        )
-        tokenizer = AutoTokenizer.from_pretrained(module)
-
-        multi_line_print(
-            f"Training: Model loaded {config.task_type.value}: {PrintFormats.BOLD}{module.__class__.__name__ if isinstance(module, PreTrainedModel) else module}{PrintFormats.END}"
-        )
-    except:
-        multi_line_print(
-            f"Training: Couldn't load {module} model or tokenizer. Skipping."
-        )
-        model, tokenizer = None, None
-
-    return model, tokenizer
-
-
-def safe_load_inference_pipeline(
-    module: Union[str, PreTrainedModel],
-    config: HuggingfaceConfig,
-    tokenizer: Optional[PreTrainedTokenizerBase] = None,
-) -> Optional[Callable]:
-    try:
-        if tokenizer is not None:
-            loaded_pipeline = pipeline(
-                task=config.task_type.value,
-                model=module,
-                tokenizer=tokenizer,
-                device=device,
-            )
-        else:
-            loaded_pipeline = pipeline(
-                task=config.task_type.value, model=module, device=device
-            )
-
-        multi_line_print(
-            f"Inference: Pipeline loaded {config.task_type.value}: {PrintFormats.BOLD}{module.__class__.__name__ if isinstance(module, PreTrainedModel) else module}{PrintFormats.END}"
-        )
-
-    except:
-        multi_line_print(f"Inference: Couldn't load {module} pipeline. Skipping.")
-
-        loaded_pipeline = None
-
-    return loaded_pipeline
-
-
-def safe_load(
-    train: bool,
-    module: Union[str, PreTrainedModel],
-    config: HuggingfaceConfig,
-    tokenizer: Optional[PreTrainedTokenizerBase] = None,
-) -> Tuple[
-    Optional[Union[Callable, PreTrainedModel]], Optional[PreTrainedTokenizerBase]
-]:
-
-    if train and isinstance(module, str):
-        return safe_load_training_pipeline(module, config)
-    else:
-        return safe_load_inference_pipeline(module, config, tokenizer), None
 
 
 def initalize_environment_(config: HuggingfaceConfig) -> None:
     os.environ["TOKENIZERS_PARALLELISM"] = "False"
     config.training_args.hub_token = get_env("HF_HUB_TOKEN")
-
-
-def get_paths(config: HuggingfaceConfig, parent_path: str, id: str) -> dict:
-    return {
-        LoadOrigin.local: f"{Const.output_pipelines_path}/{parent_path}/{id}",
-        LoadOrigin.remote: f"{config.user_name}/{id}"
-        if not hasattr(config, "remote_name_override")
-        or config.remote_name_override is None
-        else config.remote_name_override,
-        LoadOrigin.pretrained: config.pretrained_model,
-    }
-
-
-def determine_load_order(
-    config: HuggingfaceConfig, paths: dict
-) -> List[Tuple[LoadOrigin, str]]:
-    if (
-        hasattr(config, "preferred_load_origin")
-        and config.preferred_load_origin is not None
-    ):
-        load_order = [
-            (
-                config.preferred_load_origin,
-                paths[config.preferred_load_origin],
-            )
-        ]
-    else:
-        load_order = paths.items()
-
-    return load_order
 
 
 class HuggingfaceModel(Model):
